e-test-server-utils: Destroy dynamic closures
Adds a callback to the closure structure that describes
its concept of self-destruction.

Conflicts:

	tests/test-server-utils/e-test-server-utils.h

+++ tests/test-server-utils/e-test-server-utils.c
@@ -359,8 +359,12 @@ e_test_server_utils_teardown (ETestServerFixture *fixture,
 	}
 
 	/* Cleanup work directory */
-	if (!test_installed_services ())
+	if (!closure->keep_work_directory && !test_installed_services ())
 		delete_work_directory ();
+
+	/* Destroy dynamically allocated closure */
+	if (closure->destroy_closure_func)
+		closure->destroy_closure_func (closure);
 }
 
 gint
